+++ next-docs/tailwind.config.js
@@ -3,6 +3,10 @@ module.exports = {
     './public/**/*.{js,ts,jsx,tsx}',
     './pages/**/*.{js,ts,jsx,tsx}',
     './components/**/*.{js,ts,jsx,tsx}',
+    './node_modules/@heathmont/moon-core-tw/**/*.{js,ts,jsx,tsx}',
+  ],
+  presets: [
+    require('@heathmont/moon-core-tw/src/private/presets/ds-moon-preset'),
   ],
   theme: {
     extend: {
@@ -25,94 +29,7 @@ module.exports = {
         developer: 'rgba(149, 241, 213, 0.9)',
         transparent: 'transparent',
         current: 'currentColor',
-        piccolo: {
-          120: 'var(--piccolo-120)',
-          DEFAULT: 'var(--piccolo)',
-          80: 'var(--piccolo-80)',
-        },
-        hit: {
-          120: 'var(--hit-120)',
-          DEFAULT: 'var(--hit)',
-          80: 'var(--hit-80)',
-        },
-        beerus: {
-          100: 'var(--beerus)',
-        },
-        goku: {
-          DEFAULT: 'var(--goku)',
-          80: 'var(--goku-80)',
-          40: 'var(--goku-40)',
-          10: 'var(--goku-10)',
-        },
-        gohan: {
-          DEFAULT: 'var(--gohan)',
-          80: 'var(--gohan-80)',
-          40: 'var(--gohan-40)',
-          10: 'var(--gohan-10)',
-        },
-        bulma: {
-          DEFAULT: 'var(--bulma)',
-        },
-        trunks: {
-          DEFAULT: 'var(--trunks)',
-        },
-        goten: {
-          DEFAULT: 'var(--goten)',
-        },
-        popo: {
-          DEFAULT: 'var(--popo)',
-        },
-        // support colours
-        krillin: {
-          DEFAULT: 'var(--krillin)',
-          60: 'var(--krillin-60)',
-          10: 'var(--krillin-10)',
-        },
-        chiChi: {
-          DEFAULT: 'var(--chiChi)',
-          60: 'var(--chiChi-60)',
-          10: 'var(--chiChi-10)',
-        },
-        roshi: {
-          DEFAULT: 'var(--roshi)',
-          60: 'var(--roshi-60)',
-          10: 'var(--roshi-10)',
-        },
-        dodoria: {
-          DEFAULT: 'var(--dodoria)',
-          60: 'var(--dodoria-60)',
-          10: 'var(--dodoria-10)',
-        },
-        cell: {
-          DEFAULT: 'var(--cell)',
-          60: 'var(--cell-60)',
-          10: 'var(--cell-10)',
-        },
-        raditz: {
-          DEFAULT: 'var(--raditz)',
-          60: 'var(--raditz-60)',
-          10: 'var(--raditz-10)',
-        },
-        whis: {
-          DEFAULT: 'var(--whis)',
-          60: 'var(--whis-60)',
-          10: 'var(--whis-10)',
-        },
-        frieza: {
-          DEFAULT: 'var(--frieza)',
-          60: 'var(--frieza-60)',
-          10: 'var(--frieza-10)',
-        },
-        nappa: {
-          DEFAULT: 'var(--nappa)',
-          60: 'var(--nappa-60)',
-          10: 'var(--nappa-10)',
-        },
       },
     },
   },
-  variants: {
-    extend: {},
-  },
-  plugins: [require('tailwindcss-rtl')],
 };