+++ workspaces/core/src/private/utils/classnames.ts
@@ -1,0 +1,5 @@
+function classNames(...classes: unknown[]): string {
+  return classes.filter(Boolean).join(' ');
+}
+
+export default classNames;